DRY workspace version resolution check
